+++ src/main/java/com/collab/clicker/Upgrade.java
@@ -3,5 +3,5 @@ package com.collab.clicker;
 public interface Upgrade {
     void update(Miner miner);
     String getDescription();
-    long getPrice();
+    double getPrice();
 }
